1. update the fordfulkerson alg such that the cost of the max flow is cacluated as sum(cpath_i*bw_i)/sum(bw_i), where cpath_i is the sum of the cost of all edges in one path and bw_i is the min of the bw of all edges in one path
2. deal with the error situation, when a error happened, not only an 
error msg is outputed but also exit the program

3. update the construction of virtual graph, where the case when 
multiple edges exist between two nodes is taken care

4. update the construction of inter-as graph, where m_left now is 
defined the as where the start node resides in, and m_right is the as 
where the end node resides in

5. deal with the memory release for main-sdni scenario, which delete the 
commodity memory and allocation memory

6 revise updateTopoTableNaive() such that multiple paths with same 
source, next, and dst can exist in the topo table

7 revise updateAdvertiseTable() such that the operation is more concise

8 add a new file which can generate the topology where every two 
neighboring nodes contain two links between them, and the bw and weight 
of these two links are same.

9 track the table sizes (including topo table and advertisement table)

10 revise naive_TE() such that s1->s2->n2->d2 can be advertised to the 
neighbor as of s1.